Removing Emacs multimode header
This now causes a conflict with noweb in Emacs 24, resulting in an error "no web mode removed" which comes from ESS lisp code. No longer needed and perhaps has not been for some time. Don't recall the etiology of including this from ESS use years ago, but at some point it was needed.